requirements.txt
feedparser==6.0.11
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
Jinja2==3.1.4
pytz==2024.1
beautifulsoup4==4.12.3
lxml==5.1.0
